moves all text normalizer code to a new "normalizer" sub-package
