+++ src/main/java/com/acciente/oacc/normalizer/TextNormalizer.java
@@ -16,7 +16,11 @@
  * permissions and limitations under the License.
  */
 
-package com.acciente.oacc.encryptor;
+package com.acciente.oacc.normalizer;
+
+import com.acciente.oacc.normalizer.icu4j.ICU4JTextNormalizerForICU4JBelowVersion4dot6;
+import com.acciente.oacc.normalizer.icu4j.ICU4JTextNormalizerForICU4JVersion4dot6AndHigher;
+import com.acciente.oacc.normalizer.jdk.JDKTextNormalizer;
 
 /**
  * Normalizes Unicode text to handle characters that have more than one canonically equivalent representation.
