+++ src/main/java/com/acciente/oacc/normalizer/icu4j/ICU4JTextNormalizerForICU4JVersion4dot6AndHigher.java
@@ -16,8 +16,9 @@
  * permissions and limitations under the License.
  */
 
-package com.acciente.oacc.encryptor;
+package com.acciente.oacc.normalizer.icu4j;
 
+import com.acciente.oacc.normalizer.TextNormalizer;
 import com.ibm.icu.text.Normalizer2;
 
 import java.nio.CharBuffer;
