+++ src/test/java/com/acciente/oacc/normalizer/TextNormalizerInstanceTest.java
@@ -15,8 +15,9 @@
  * See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package com.acciente.oacc.encryptor;
+package com.acciente.oacc.normalizer;
 
+import com.acciente.oacc.normalizer.icu4j.ICU4JTextNormalizerForICU4JVersion4dot6AndHigher;
 import org.junit.Test;
 
 import static org.hamcrest.core.IsEqual.equalTo;
