+++ src/test/java/com/acciente/oacc/normalizer/TextNormalizerTest.java
@@ -16,8 +16,11 @@
  * permissions and limitations under the License.
  */
 
-package com.acciente.oacc.encryptor;
+package com.acciente.oacc.normalizer;
 
+import com.acciente.oacc.normalizer.icu4j.ICU4JTextNormalizerForICU4JBelowVersion4dot6;
+import com.acciente.oacc.normalizer.icu4j.ICU4JTextNormalizerForICU4JVersion4dot6AndHigher;
+import com.acciente.oacc.normalizer.jdk.JDKTextNormalizer;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
