+++ src/test/java/com/acciente/oacc/normalizer/icu4j/ICU4JLegacyNormalizerDestBufferWorstCaseExpansionTest.java
@@ -16,24 +16,20 @@
  * permissions and limitations under the License.
  */
 
-package com.acciente.oacc.encryptor;
+package com.acciente.oacc.normalizer.icu4j;
 
-import com.ibm.icu.text.Normalizer2;
-import org.junit.Before;
+import com.ibm.icu.text.Normalizer;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameter;
 import org.junit.runners.Parameterized.Parameters;
 
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.lessThanOrEqualTo;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 
 @RunWith(Parameterized.class)
-public class ICU4JNormalizer2DestBufferWorstCaseExpansionTest {
-   private Normalizer2 normalizer;
+public class ICU4JLegacyNormalizerDestBufferWorstCaseExpansionTest {
 
    @Parameters
    public static Object[] data() {
@@ -55,31 +51,23 @@ public class ICU4JNormalizer2DestBufferWorstCaseExpansionTest {
    @Parameter
    public String src;
 
-   @Before
-   public void setUp() throws Exception {
-      normalizer = Normalizer2Factory.getNFCInstance();
-   }
-
    @Test
    public void testExpansion() throws Exception {
       final int expectedMaxExpansionSize = 3 * src.length();
 
       // allocate the destination to be 3x of the source length
-      StringBuilder dest = new StringBuilder(expectedMaxExpansionSize);
-      assertEquals(expectedMaxExpansionSize, dest.capacity());
+      char[] dest = new char[expectedMaxExpansionSize];
 
       // normalize the text
-      normalizer.normalize(src, dest);
-      printStats(src, dest);
-      assertThat(dest.length(), lessThanOrEqualTo(expectedMaxExpansionSize));
-      assertThat(dest.capacity(), equalTo(expectedMaxExpansionSize));
+      final int actualDestLen = Normalizer.normalize(src.toCharArray(), dest, Normalizer.NFC, 0);
+      printStats(src, actualDestLen);
+      assertThat(actualDestLen, lessThanOrEqualTo(expectedMaxExpansionSize));
    }
 
-   private void printStats(String src, StringBuilder dest) {
+   private void printStats(String src, int destLen) {
       System.out.printf("\nsrc=%s", src);
-      System.out.printf("\ndest=%s", dest);
+      System.out.printf("\ndest=%s", destLen);
       System.out.printf("\nsrc.length()=%s", src.length());
-      System.out.printf("\ndest.length()=%s", dest.length());
-      System.out.printf("\ndest.capacity()=%s", dest.capacity());
+      System.out.printf("\ndest.length()=%s", destLen);
    }
 }
